# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ra Challenge API",
        "title": "Ra Challenge API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Thiago Menezes",
            "email": "thg.mnzs@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/complains": {
            "get": {
                "description": "Get details of all complains",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Complains"
                ],
                "summary": "Get details of all complains",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ComplainOut"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new complain with the input paylod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Complains"
                ],
                "summary": "Create a new complain",
                "parameters": [
                    {
                        "description": "Create complain",
                        "name": "complain",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ComplainIn"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/complains/{parameter}": {
            "get": {
                "description": "List complains by company",
                "summary": "Get complains by company",
                "tags": [
                    "Complains"
                ],
                "deprecated": false,
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "parameter",
                        "in": "path",
                        "description": "Name company",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succesful operation"
                    },
                    "500": {
                        "description": "error internal server"
                    }
                }
            },
            "put": {
                "description": "Update complain",
                "summary": "Update Complain",
                "tags": [
                    "Complains"
                ],
                "deprecated": false,
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "parameter",
                        "in": "path",
                        "description": "ID of the complain",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "description": "Request body",
                        "schema": {
                            "$ref": "#/definitions/ComplainModelUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "bad request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "error internal server"
                    }
                }
            },
            "delete": {
                "description": "Delete complain by ID",
                "summary": "Delete complain",
                "tags": [
                    "Complains"
                ],
                "deprecated": false,
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "parameter",
                        "in": "path",
                        "description": "ID of the complain",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "error internal server"
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Company": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.ComplainIn": {
            "type": "object",
            "required": [
                "company",
                "description",
                "title"
            ],
            "properties": {
                "company": {
                    "type": "object",
                    "$ref": "#/definitions/model.Company"
                },
                "description": {
                    "type": "string"
                },
                "locale": {
                    "type": "object",
                    "$ref": "#/definitions/model.Locale"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.ComplainOut": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "object",
                    "$ref": "#/definitions/model.Company"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "locale": {
                    "type": "object",
                    "$ref": "#/definitions/model.Locale"
                },
                "countPageViews": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.Locale": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        }
    }
}